+++ trixnity-messenger/src/jvmMain/kotlin/de/connect2x/trixnity/messenger/CreateRepositoriesModule.jvm.kt
@@ -1,7 +1,7 @@
 package de.connect2x.trixnity.messenger
 
-import androidx.room.Room
-import androidx.room.RoomDatabase
+import androidx.room3.Room
+import androidx.room3.RoomDatabase
 import org.koin.core.scope.Scope
 
 internal actual inline fun <reified T : RoomDatabase> Scope.roomDatabaseBuilder(name: String): RoomDatabase.Builder<T> =